[refactor] to adapt to ledis

+++ bench_test.go
@@ -190,7 +190,7 @@ func BenchmarkZAdd(b *testing.B) {
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
 			err := client.ZAdd("key", &redis.Z{
-				Score:  float64(1),
+				Score:  1,
 				Member: "hello",
 			}).Err()
 			if err != nil {
